feat[pipeline]: add Docker image builder script
scripts/build_docker_image.sh, when run without arguments, creates a
Docker image with the current Git worktree contents, including
uncommitted and/or untracked files, but excluding .gitignored files.
The image is then tagged karma3labs/fcgraph-pipeline:latest.

If given -p (production mode), it uses only the committed contents of
the HEAD commit, not the current worktree contents.

The repo:tag (karma3labs/fcgraph-pipeline:latest) can be set with the -t
option, for compatibility with "docker build".  As a shortcut, if the -t
argument starts with a colon, the default repo will be prepended, e.g.,
-t :mytag is equivalent to -t karma3labs/fcgraph-pipeline:mytag.

Extra non-option arguments (that is, after the first "--") are passed
down as extra "docker build" options.